# SelectExpression.VisitChildren optimizations