Profiling the scenario in #17455 showed that SelectExpression.VisitChildren was a major compilation perf hotspot (own time, not including child calls). This PR changes the logic to only allocate new data structures and copy if a Visit actually returned a different result.

This reduced the time for #17455 from 22 seconds to 15 on my machine. Memory allocated notably goes down from 61.86MB to 12.17MB.

Also added a compilation micro-benchmark which does three joins, here are the results. Note that #17455 includes 6 results and so the difference is considerably more significant.

### Before
```
BenchmarkDotNet=v0.11.3, OS=ubuntu 19.04
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview8-013656
  [Host] : .NET Core 3.0.0-rc1-19430-09 (CoreCLR 4.700.19.43003, CoreFX 4.700.19.42010), 64bit RyuJIT

Toolchain=InProcessToolchain  

        Method |     Mean |    Error |   StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------------- |---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
 MultipleJoins | 215.8 ms | 4.067 ms | 3.605 ms |  15000.0000 |   1000.0000 |           - |            61.86 MB |
```

### After
```
BenchmarkDotNet=v0.11.3, OS=ubuntu 19.04
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview8-013656
  [Host] : .NET Core 3.0.0-rc1-19430-09 (CoreCLR 4.700.19.43003, CoreFX 4.700.19.42010), 64bit RyuJIT

Toolchain=InProcessToolchain  

        Method |     Mean |    Error |   StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------------- |---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
 MultipleJoins | 173.2 ms | 3.538 ms | 3.309 ms |   2666.6667 |    666.6667 |           - |            12.17 MB |
```